Restructure benchmarks layout
