+++ benchmarks/Main.hs
@@ -2,127 +2,12 @@ module Main where
 
 import Protolude
 
-import BinaryField
 import Criterion.Main
-import ExtensionField
-import GaloisField
-import GHC.Base
-import PrimeField
 
-type Fq = PrimeField 21888242871839275222246405745257275088696311157297823662689037894645226208583
-
-fq :: Fq
-fq = 5216004179354450092383934373463611881445186046129513844852096383579774061693
-
-fq' :: Fq
-fq' = 10757805228921058098980668000791497318123219899766237205512608761387909753942
-
-data Pu
-instance IrreducibleMonic Fq Pu where
-  split _ = x ^ (2 :: Int) + 1
-type Fq2 = ExtensionField Fq Pu
-
-data Pv
-instance IrreducibleMonic Fq2 Pv where
-  split _ = x ^ (3 :: Int) - 9 - t x
-type Fq6 = ExtensionField Fq2 Pv
-
-data Pw
-instance IrreducibleMonic Fq6 Pw where
-  split _ = x ^ (2 :: Int) - t x
-type Fq12 = ExtensionField Fq6 Pw
-
-fq12 :: Fq12
-fq12 = fromList
-  [ fromList
-    [ fromList
-      [ 4025484419428246835913352650763180341703148406593523188761836807196412398582
-      , 5087667423921547416057913184603782240965080921431854177822601074227980319916
-      ]
-    , fromList
-      [ 8868355606921194740459469119392835913522089996670570126495590065213716724895
-      , 12102922015173003259571598121107256676524158824223867520503152166796819430680
-      ]
-    , fromList
-      [ 92336131326695228787620679552727214674825150151172467042221065081506740785
-      , 5482141053831906120660063289735740072497978400199436576451083698548025220729
-      ]
-    ]
-  , fromList
-    [ fromList
-      [ 7642691434343136168639899684817459509291669149586986497725240920715691142493
-      , 1211355239100959901694672926661748059183573115580181831221700974591509515378
-      ]
-    , fromList
-      [ 20725578899076721876257429467489710434807801418821512117896292558010284413176
-      , 17642016461759614884877567642064231230128683506116557502360384546280794322728
-      ]
-    , fromList
-      [ 17449282511578147452934743657918270744212677919657988500433959352763226500950
-      , 1205855382909824928004884982625565310515751070464736233368671939944606335817
-      ]
-    ]
-  ]
-
-fq12' :: Fq12
-fq12' = fromList
-  [ fromList
-    [ fromList
-      [ 495492586688946756331205475947141303903957329539236899715542920513774223311
-      , 9283314577619389303419433707421707208215462819919253486023883680690371740600
-      ]
-    , fromList
-      [ 11142072730721162663710262820927009044232748085260948776285443777221023820448
-      , 1275691922864139043351956162286567343365697673070760209966772441869205291758
-      ]
-    , fromList
-      [ 20007029371545157738471875537558122753684185825574273033359718514421878893242
-      , 9839139739201376418106411333971304469387172772449235880774992683057627654905
-      ]
-    ]
-  , fromList
-    [ fromList
-      [ 9503058454919356208294350412959497499007919434690988218543143506584310390240
-      , 19236630380322614936323642336645412102299542253751028194541390082750834966816
-      ]
-    , fromList
-      [ 18019769232924676175188431592335242333439728011993142930089933693043738917983
-      , 11549213142100201239212924317641009159759841794532519457441596987622070613872
-      ]
-    , fromList
-      [ 9656683724785441232932664175488314398614795173462019188529258009817332577664
-      , 20666848762667934776817320505559846916719041700736383328805334359135638079015
-      ]
-    ]
-  ]
-
-type F2m = BinaryField 0x80000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000425
-
-f2m :: F2m
-f2m = 0x303001d34b856296c16c0d40d3cd7750a93d1d2955fa80aa5f40fc8db7b2abdbde53950f4c0d293cdd711a35b67fb1499ae60038614f1394abfa3b4c850d927e1e7769c8eec2d19
-
-f2m' :: F2m
-f2m' = 0x37bf27342da639b6dccfffeb73d69d78c6c27a6009cbbca1980f8533921e8a684423e43bab08a576291af8f461bb2a8b3531d2f0485c19b16e2f1516e23dd3c1a4827af1b8ac15b
-
-benchmark :: GaloisField k => String -> k -> k -> Benchmark
-benchmark s a b = bgroup s
-  [ bench "Addition" $
-    whnf (uncurry (+)) (a, b)
-  , bench "Multiplication" $
-    whnf (uncurry (*)) (a, b)
-  , bench "Negation" $
-    whnf negate a
-  , bench "Subtraction" $
-    whnf (uncurry (-)) (a, b)
-  , bench "Inversion" $
-    whnf recip a
-  , bench "Division" $
-    whnf (uncurry (/)) (a, b)
-  ]
+import BinaryFieldBenchmarks
+import ExtensionFieldBenchmarks
+import PrimeFieldBenchmarks
 
 main :: IO ()
 main = defaultMain
-  [ benchmark "PrimeField Fq" fq fq'
-  , benchmark "ExtensionField Fq12" fq12 fq12'
-  , benchmark "BinaryField F2m" f2m f2m'
-  ]
+  [benchmarkBinaryField, benchmarkExtensionField, benchmarkPrimeField]
